requirements/base.txt
django==2.0.3
django-ckeditor==5.4.0
django-mptt==0.9.0
django-compressor==2.2
django-environ==0.4.4

psycopg2==2.7.4
pillow==5.0.0
sorl-thumbnail==12.4.1
python-slugify==1.2.5
beautifulsoup4==4.6.0
